requirements.txt
pyyaml
psutil
rich
pydantic